Changes prepoceanobs to using newly refactored BUFR converters (#1352)
Just like the title. This mainly changes the converters used to the new
ones that use yaml files, the templates for which are included, along
with a little script that does most of the work of converting the
earlier JSON templates to yaml. Also removes from `ocean.py` some
methods for plotting, but can't go into global-workflow because of lack
of the necessary python modules in that environment.