Fixes gridgen problem and localization for LETKF (#1372)
Adds the `gridgen.yaml ` back after being removed to` jcb-gdas`, to
allow LETKF to work with a fix within GDASApp. This is a temporary fix
until the LETKF task can be converted to using the jcb yamls.

Also adds `obs localization` to remaining obs config files - this will
have to likewise be carried over to whatever is being used with the obs
list.