Add variable rescaling for aerosol species (#1339)
This is a draft PR for additions to chem_diagb.h that could allow for a
variable rescaling (rs) for each aerosol species. It does not yet
include a way to read the values in from a yaml or config file. The it
compiles and runs, but I have not yet tested it using 3dvar. And it
still has lots of print statements used for testing. It runs in about 40
seconds at C96, and is probably slowed down by the print statements. Any
comments will be helpful.

Also, note the if statement. I was unsure of the dimensions of an array,
so I put in a test. I'll remove it along with the print statements.