Add more CI to GW-CI (#1365)
This PR is a companion to GW PR
[#3120](https://github.com/NOAA-EMC/global-workflow/pull/3120).

It does a couple things:

1. 5 GW CI tests are added/extended as CTests in GDASApp, running
through to the fcst jobs in the first full-cycle. These CI tests are:
```C96C48_hybatmDA```, ```C96C48_ufs_hybatmDA```,
```C96C48_hybatmaerosnowDA```, ```C48mx500_3DVarAOWCDA```, and
```C48mx500_hybAOWCDA```.
2. Test references are added for ```C96C48_ufs_hybatmDA``` and
```C48mx500_3DVarAOWCDA```, so that we're actually testing the output.
3. These CTests are turned on by default in a workflow build, rather
than having to mess with the ```CMakeCache.txt``` file and re-running
make. This will allow us to use these tests in nightly testing.
4. ```test/gw-ci/CMakeLists.txt``` is refactored quite a bit.
5. There are 89 CTests, but for 5 CI tests, but I added task
dependencies, so they can be run in parallel.

The primary motivation for this PR is that we can run CI for our nightly
testing of GDASApp. Also, anyone with a PR can easily do CI testing
through CTests.

---------

Co-authored-by: Russ-Treadon-NOAA <Russ.Treadon@noaa.gov>