Make sure gfs_* DA jobs are not regression tests (#1390)
This PR addresses the bug @RussTreadon-NOAA found that
```gfs_atmanlvar``` was being run as a regression test and using the
same test reference as ```gdas_atmanlvar``` in GW PR
[#3120](https://github.com/NOAA-EMC/global-workflow/pull/3120).

See
https://github.com/NOAA-EMC/global-workflow/pull/3120#issuecomment-2501783018

I've moved all activation of testing mode in JCB out of the JCB base
YAMLs and into the JCB algorithm YAMLs. I test the ```RUN``` variables
to make sure it's not equal to ```gfs```.

I re-ran all the regression tests, and they all passed.